+++ example/linux/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,13 @@
+//
+//  Generated file. Do not edit.
+//
+
+#include "generated_plugin_registrant.h"
+
+#include <flutter_barcode_sdk/flutter_barcode_sdk_plugin.h>
+
+void fl_register_plugins(FlPluginRegistry* registry) {
+  g_autoptr(FlPluginRegistrar) flutter_barcode_sdk_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "FlutterBarcodeSdkPlugin");
+  flutter_barcode_sdk_plugin_register_with_registrar(flutter_barcode_sdk_registrar);
+}