+++ lib/flutter_barcode_sdk_web.dart
@@ -6,10 +6,13 @@ import 'dart:async';
 import 'dart:html' as html show window;
 
 import 'package:flutter/services.dart';
+import 'package:flutter_barcode_sdk/barcode_manager.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
 /// A web implementation of the FlutterBarcodeSdk plugin.
 class FlutterBarcodeSdkWeb {
+  BarcodeManager _barcodeManager = BarcodeManager();
+
   static void registerWith(Registrar registrar) {
     final MethodChannel channel = MethodChannel(
       'flutter_barcode_sdk',
@@ -28,11 +31,16 @@ class FlutterBarcodeSdkWeb {
     switch (call.method) {
       case 'getPlatformVersion':
         return getPlatformVersion();
+      case 'decodeFile':
+        return decodeFile(call.arguments['filename']);
+      case 'decodeVideo':
+        decodeVideo();
         break;
       default:
         throw PlatformException(
           code: 'Unimplemented',
-          details: 'flutter_barcode_sdk for web doesn\'t implement \'${call.method}\'',
+          details:
+              'flutter_barcode_sdk for web doesn\'t implement \'${call.method}\'',
         );
     }
   }
@@ -41,5 +49,15 @@ class FlutterBarcodeSdkWeb {
   Future<String> getPlatformVersion() {
     final version = html.window.navigator.userAgent;
     return Future.value(version);
+  }
+
+  /// Decode barcodes from an image file.
+  Future<List<Map<dynamic, dynamic>>> decodeFile(String file) async {
+    return _barcodeManager.decodeFile(file);
+  }
+
+  /// Decode barcodes from real-time video stream.
+  Future<void> decodeVideo() async {
+    _barcodeManager.decodeVideo();
   }
 }