+++ lib/global.dart
@@ -1,0 +1,2 @@
+/// Callback for barcode scanner.
+Function globalCallback = () => {};